refactor(frontend): modularize all major pages
- Extract business logic to custom hooks for EditProfile, Login, ForgotPassword
- Extract presentational forms to components (EditProfileForm, LoginForm, ForgotPasswordForm)
- Refactor pages to use new hooks and components
- Ensure SOLID, DRY, and reviewer-friendly structure
- No logic changes, only modularization and maintainability improvements
